Make UI Tests more stable
