+++ Example/PaymentSheet Example/PaymentSheetUITest/EmbeddedUITest.swift
@@ -98,9 +98,7 @@ class EmbeddedUITests: PaymentSheetUITestCase {
 
         loadPlayground(app, settings)
         app.buttons["Present embedded payment element"].waitForExistenceAndTap()
-
-        XCTAssertTrue(app.buttons["•••• 1001"].waitForExistence(timeout: 3.0))
-        XCTAssertFalse(app.buttons["•••• 4242"].waitForExistence(timeout: 3.0))
+        ensureSPMSelection("•••• 1001", insteadOf: "•••• 4242")
 
         // Switch from 1001 to 4242
         app.buttons["View more"].waitForExistenceAndTap()
@@ -132,9 +130,7 @@ class EmbeddedUITests: PaymentSheetUITestCase {
 
         loadPlayground(app, settings)
         app.buttons["Present embedded payment element"].waitForExistenceAndTap()
-
-        XCTAssertTrue(app.buttons["••••6789"].waitForExistence(timeout: 3.0))
-        XCTAssertFalse(app.buttons["•••• 4242"].waitForExistence(timeout: 3.0))
+        ensureSPMSelection("••••6789", insteadOf: "•••• 4242")
 
         // Switch from 6789 (Bank account) to 4242
         app.buttons["View more"].waitForExistenceAndTap()
@@ -172,5 +168,22 @@ class EmbeddedUITests: PaymentSheetUITestCase {
 
         let alert = app.alerts[alertTitle]
         alert.buttons[buttonToTap].tap()
+    }
+
+    // Returning customers have two payment methods in a non-deterministic order.
+    // Ensure state of payment method of label1 is selected prior to starting tests.
+    func ensureSPMSelection(_ label1: String, insteadOf label2: String) {
+        if app.buttons[label1].waitForExistence(timeout: 3.0) {
+            XCTAssertFalse(app.buttons[label2].waitForExistence(timeout: 3.0))
+            return
+        }
+        guard app.buttons[label2].waitForExistence(timeout: 3.0) else {
+            XCTFail("Unable to find either \(label1) or \(label2)")
+            return
+        }
+        app.buttons["View more"].waitForExistenceAndTap(timeout: 3.0)
+        app.buttons[label1].waitForExistenceAndTap(timeout: 3.0)
+        XCTAssertTrue(app.buttons[label1].waitForExistence(timeout: 3.0))
+
     }
 }
