+++ Testers/IntegrationTester/IntegrationTesterUITests/XCUITest+Utilities.swift
@@ -1,0 +1,1 @@
+../../../Example/PaymentSheet Example/PaymentSheetUITest/XCUITest+Utilities.swift
